requirements.txt
google-genai
regex
YouTubeTranscriptApi
